+++ Content.IntegrationTests/Tests/Construction/ConstructionPrototypeTest.cs
@@ -1,9 +1,8 @@
 using System.Numerics;
-using Content.Server.Construction.Components;
+using Content.Shared.Construction.Components;
 using Content.Shared.Construction.Prototypes;
 using Robust.Shared.GameObjects;
 using Robust.Shared.Map;
-using Robust.Shared.Maths;
 using Robust.Shared.Prototypes;
 
 namespace Content.IntegrationTests.Tests.Construction
