+++ Content.IntegrationTests/Tests/Interaction/InteractionTest.Helpers.cs
@@ -6,11 +6,11 @@ using System.Numerics;
 using System.Reflection;
 using Content.Client.Construction;
 using Content.Server.Atmos.EntitySystems;
-using Content.Server.Construction.Components;
 using Content.Server.Gravity;
 using Content.Server.Power.Components;
 using Content.Shared.Atmos;
 using Content.Shared.CombatMode;
+using Content.Shared.Construction.Components;
 using Content.Shared.Construction.Prototypes;
 using Content.Shared.Gravity;
 using Content.Shared.Item;
