+++ Content.Server/Construction/ConstructionSystem.Computer.cs
@@ -1,6 +1,7 @@
 using Content.Server.Construction.Components;
 using Content.Server.Power.Components;
 using Content.Shared.Computer;
+using Content.Shared.Construction.Components;
 using Content.Shared.Power;
 using Robust.Shared.Containers;
