+++ Content.Server/Construction/ConstructionSystem.Interactions.cs
@@ -456,26 +456,6 @@ namespace Content.Server.Construction
         }
 
         /// <summary>
-        ///     Performs a number of <see cref="IGraphAction"/>s for a given entity, with an optional user entity.
-        /// </summary>
-        /// <param name="uid">The entity to perform the actions on.</param>
-        /// <param name="userUid">An optional user entity to pass into the actions.</param>
-        /// <param name="actions">The actions to perform.</param>
-        /// <remarks>This method checks whether the given target entity exists before performing any actions.
-        ///          If the entity is deleted by an action, it will short-circuit and stop performing the rest of actions.</remarks>
-        public void PerformActions(EntityUid uid, EntityUid? userUid, IEnumerable<IGraphAction> actions)
-        {
-            foreach (var action in actions)
-            {
-                // If an action deletes the entity, we stop performing the rest of actions.
-                if (!Exists(uid))
-                    break;
-
-                action.PerformAction(uid, userUid, EntityManager);
-            }
-        }
-
-        /// <summary>
         ///     Resets the current construction edge status on an entity.
         /// </summary>
         /// <param name="uid">The target entity.</param>
