+++ Content.Server/Construction/ConstructionSystem.cs
@@ -1,7 +1,8 @@
-using Content.Server.Construction.Components;
 using Content.Server.Stack;
 using Content.Shared.Construction;
+using Content.Shared.Construction.Components;
 using Content.Shared.DoAfter;
+using Content.Shared.Verbs;
 using JetBrains.Annotations;
 using Robust.Server.Containers;
 using Robust.Shared.Random;
@@ -27,13 +28,17 @@ namespace Content.Server.Construction
 
             InitializeComputer();
             InitializeGraphs();
-            InitializeGuided();
             InitializeInteractions();
             InitializeInitial();
             InitializeMachines();
 
             SubscribeLocalEvent<ConstructionComponent, ComponentInit>(OnConstructionInit);
             SubscribeLocalEvent<ConstructionComponent, ComponentStartup>(OnConstructionStartup);
+        }
+
+        public override void BeginDeconstructionGuide(EntityUid uid, ConstructionComponent component, GetVerbsEvent<Verb> args)
+        {
+            SetPathfindingTarget(uid, component.DeconstructionNode, component);
         }
 
         private void OnConstructionInit(Entity<ConstructionComponent> ent, ref ComponentInit args)
